fix: OptionRepository을 <Option, Long>으로 변경

+++ src/main/java/gift/option/repository/OptionRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OptionRepository extends JpaRepository<Option, Integer> {}
+public interface OptionRepository extends JpaRepository<Option, Long> {}
